190430 : 유도 + 포물선 프로젝타일 구현 (다듬어야함)

+++ Source/work2/Tower.h
@@ -74,7 +74,7 @@ public:
 	virtual float GetTowerAttackSpd();
 
 	UFUNCTION(BlueprintCallable, BlueprintPure, Category = "UI|HUD")
-	virtual float GetTowerAttackDmg();
+	virtual int32 GetTowerAttackDmg();
 
 	UFUNCTION(BlueprintImplementableEvent, BlueprintCallable)
 	void ResponseButtonEvent(int iNum);
